+++ sdk/src/wallet/message_interface/message_handler.rs
@@ -38,7 +38,7 @@ use crate::{
             operations::transaction::{
                 high_level::minting::mint_native_token::MintTokenTransactionDto, TransactionOptions,
             },
-            types::{AccountIdentifier, BalanceDto, TransactionDto},
+            types::{AccountIdentifier, TransactionDto},
             OutputDataDto,
         },
         message_interface::{
@@ -727,7 +727,7 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::GetBalance => Ok(Response::Balance(BalanceDto::from(&account.balance().await?))),
+            AccountMethod::GetBalance => Ok(Response::Balance(account.balance().await?)),
             AccountMethod::PrepareOutput {
                 params,
                 transaction_options,
@@ -781,9 +781,7 @@ impl WalletMessageHandler {
                 })
                 .await
             }
-            AccountMethod::SyncAccount { options } => {
-                Ok(Response::Balance(BalanceDto::from(&account.sync(options).await?)))
-            }
+            AccountMethod::SyncAccount { options } => Ok(Response::Balance(account.sync(options).await?)),
             AccountMethod::SendAmount { params, options } => {
                 convert_async_panics(|| async {
                     let transaction = account